requirements.txt
torch
transformers
accelerate
pillow
pytorch-lightning